refactor(xslate): extract analysis/spread/memo/props domains
Mirror the Mojo adapter's remaining domain extractions (issue #2018 track
D) for the Text::Xslate (Kolon) adapter, each behind a narrow context seam
where it needs adapter callbacks. Behaviour unchanged — conformance
byte-identical (full suite 353 pass / 0 fail).

  analysis/component-tree.ts  hasClientInteractivity,
                              collectImportedLoopChildComponentErrors
  spread/spread-codegen.ts    conditional-spread / object-literal → Kolon
                              hashref lowering (XslateSpreadContext seam)
  memo/seed.ts                in-template derived-memo / context seeding
                              (XslateMemoContext seam)
  props/prop-classes.ts       per-compile prop classification sets

`type/` is intentionally absent (dynamically-typed template target, no
struct/type codegen), matching the Mojo adapter.
